fixed review notifications
