+++ Admin-app/src/components/PatientProfile/SymptomsRow.js
@@ -8,6 +8,9 @@ import Checkbox from "@mui/material/Checkbox";
 import { useState } from "react";
 import { useParams } from "react-router-dom";
 import { setReviewed } from "../../backend/firebasePatientUtilities";
+import { doc, serverTimestamp, addDoc, collection } from "firebase/firestore";
+
+import { db } from "../../backend/firebase";
 
 /**
  * SymptomsRow function works for setting & displaying the synptoms row of the patient
@@ -22,7 +25,19 @@ function SymptomsRow(props) {
     setReviewed(key, docID).then((newReviewedValue) =>
       setReviewedStatus(newReviewedValue)
     );
+    addStatusReviewedNotif();
   }
+
+  // This function will add notifications to the client's doc if status is reviewed
+  const addStatusReviewedNotif = async () => {
+    const clientRef = doc(db, `Client/${key}`);
+    const notifRef = collection(clientRef, "reviewNotification");
+    await addDoc(notifRef, {
+      notif: "Status Reviewed",
+      timestamp: serverTimestamp(),
+      seen: "False",
+    });
+  };
 
   return (
     <TableRow
