+++ Admin-app/src/components/PatientProfile/index.js
@@ -29,8 +29,6 @@ import {
 import DropdownStatus from "./../DropdownStatus";
 import DropdownDoctor from "./../DropdownDoctor";
 import SymptomsRow from "./SymptomsRow";
-import { db } from "../../backend/firebase";
-import { doc, serverTimestamp, addDoc, collection } from "firebase/firestore";
 
 /**
  * setAge function works for setting the age of the patient
@@ -92,13 +90,6 @@ function PatientProfile() {
     };
   }
 
-  // reviewed status with DB
-  function onReviewedClick() {
-    if (patientInfo.reviewed !== false) {
-      addStatusReviewedNotif();
-    }
-  }
-
   // priority flag with DB
   function onFlagClick() {
     togglePriorityFlag(key).then((newPatientInfo) =>
@@ -141,17 +132,6 @@ function PatientProfile() {
         console.log(err);
       });
   }, [key]);
-
-  // This function will add notifications to the client's doc if status is reviewed
-  const addStatusReviewedNotif = async () => {
-    const clientRef = doc(db, `Client/${key}`);
-    const notifRef = collection(clientRef, "reviewNotification");
-    await addDoc(notifRef, {
-      notif: "Status Reviewed",
-      timestamp: serverTimestamp(),
-      seen: "False",
-    });
-  };
 
   return (
     <Grid container spacing={2} maxWidth="lg" alignItems="flex-end">
